Evaluate using a less customized git config
I am currently evaluating how using a more default git setup would work
for me. I don't use most of the aliases any way, with the exception of
hist and s. And it seems that for the most part, the values that I
set are dated back from before the release of git 2.0, and that the
default configuration is very much in line with most of the other
configuration options that I set.