Move application shortcuts to SUPER + [letter]
I like it much better than the SUPER+SHIFT setup. I regret that I have to maintain this change separately.

At the same time, I'm enabling SUPER + hjkl to navigate windows.

I find that it is much more intuitive, but admittedly it becomes a fierce battle for the good keys.

My layout principles are now as follows:
- SUPER + <number> for namespace navigation
- SUPER + hjkl for window focus
- SUPER + c / v / x for clipboard
- SUPER + w for closing.
- SUPER + [letter] for applications

- SUPER + SHIFT + <key> is for rarely used and generally manipulative shortcuts like move to workspace, or move windows.

- SUPER + CTRL for variations over SUPER + <key> base functions, like show clipboard manager, or launch alternative version of an app.

I might not be consistent with the above in all cases, but that is the principles I tried to adhere to. I have a feeling this is not the last time I make changes to this setup.